requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cloud-sql-python-connector==1.20.0
cryptography==46.0.5
dnspython==2.8.0
frozenlist==1.8.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-pubsub==2.35.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
importlib_metadata==8.7.1
multidict==6.7.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pg8000==1.31.5
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
scramp==1.4.8
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
zipp==3.23.0
